waterproof-as-sync/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-auth==2.25.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
python-dotenv==1.0.0
